requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
streamlit>=1.30.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tiktoken>=0.5.0
